WT-11067: overhauling animateDisplay to make animations smoother
The animations were often quite jittery, and since modern browsers
have a proper JavaScript animation API with keyframes, I opted to
use that instead of the existing implementation, which relied on
adding and removing CSS classes at the appropriate time.

A lot of JSDoc type annotations were added to aid in development.