WT-10995: fix WContainerWidget's wtEncodeValue not surviving page reload
Code defined in doJavaScript is only executed once, so when we refresh
the page the wtEncodeValue function is not defined. Using
setJavaScriptMember makes sure that it is set every time. Also,
instead of relying on jsRef(), we can just use the argument that is
passed into wtEncodeValue.

We should also not throw from setFormData, and instead only log an
error, otherwise simply having no proper form values for a
WContainerWidget would cause the session to terminate, which is usually
undesirable.