JWT-11138: change <img> tags in widget gallery templates with custom resolver
In order to make the img use the correct URL regardless of the context
path, I made it so that instead of using the <img> tag, I extended
resolveString to resolve img, and run
WApplication::resolveRelativeUrl(...) on the src.

To have this function available for both WidgetGallery and
TopicTemplate, I introduced a BaseTemplate that's the base of both
templates. This also allows us to extract some other common code, e.g.
adding the tr function and setting the if:cpp and if:java conditions.