WT-10942: widget gallery: fix links to documentation
Several chart-related documentation links did not work properly due to
"Chart-" not being included.

Some other links were also fixed for JWt, by having a map from namespace
to package, and adding type and title parameters to doc-link, so we can
properly reference namespaces/packages, and have the text of the link
differ from the URL.

There's still one location where we had to manually use if:cpp/if:java,
because I'm not sure how to dynamically create a link to a function in
Doxygen documentation.