WT-10924: remove split-script configuration property
The split-script feature isn't documented anywhere, and it doesn't seem
to be entirely correct anyway?

This feature is only active when both progressive-bootstrap and
split-script are enabled.

What this feature appears to do
===============================

When a normal (not WidgetSet, nor plain HTML) session bootstraps, it
follows this pattern:

1. A page (based on Boot.html or Hybrid.html) containing the JavaScript
   from Boot.js is loaded (WebResponse::ResponseType::Page)
2. The JavaScript is loaded. This includes the JavaScript from
   jquery.js, and Wt.js, and all of the initial (visible) contents or
   the JavaScript to "upgrade" the progressive bootstrap page
   (WebResponse::ResponseType::Script)
3. A JavaScript update request is sent with the load signal
   (WebResponse::ResponseType::Update)

When split-script is enabled, the second step is split up into:

1. Loading jquery.js and Wt.js (the skeletons)
2. Upgrading the progressive bootstrap page (the rest)

This seems to be some attempt at caching part of the loaded script since
for the first part (the skeletons) the Cache-Control header is added:

Cache-Control: max-age=2592000,private

Problems with this feature
==========================

When the feature is enabled it doesn't appear to break anything, but the
cache is always defeated by the fact that the request contains the
session id and script id as parameters, which are randomly generated for
every session.

This change
===========

I removed the code in Configuration.h/C that is related to the
split-script configuration setting, and simplified all code that
referred to it:

- The skeleton parameter was used to see which part needs to be
  rendered, so all code that checked if there was a skeleton
  parameter can be removed, and we can just assume there is no
  skeleton parameter defined.
- Boot.js loaded different JavaScript depending on SPLIT_SCRIPT,
  but now we only need to keep the part if SPLIT_SCRIPT was false.
  We can also discard the callback argument of the loadScript function
  and all logic related to it.
- Since we no longer call WebRenderer::setCaching(..., true), we can
  simplify and clarify this function by no longer taking a bool argument
  and calling it addNoCacheHeaders.