WT-11240: Fix WTreeView loading
When rows are removed, WTreeView::modelRowsAboutToBeRemoved calls isExpandedRecursive
to determine if rendered tree nodes are impacted by the removal.
However, isExpandedRecursive was always returning false for the root node.
This seems odd, when the root node is always expanded. This behaviour caused a bug
because the removedHeight_ was not being set. This lead to the childrenHeight not
being updated, and the loading isn't triggered.

Switching isExpandedRecursive to true for the root index fixes the issue.