WT-11248: unset margins
Bootstrap 5 automatically sets left and right margins of the
modal-dialog to auto, and top and bottom margins to
var(--bs-modal-width).

The top and bottom margins mess with our resize code, however, and the
left and right margins cause contents to be cut off.

We can fix this for WDialog by unsetting all margins.