WT-11248: Bootstrap 5: fix scrollbar appearing on resize
When the dialog would resize, this would resize the layoutContainer
inside of it, calling wtResize with the new size of the dialog.

The existing logic took that width and height and ADDED the border
width, making the layoutContainer larger than the surrounding div.

It appears to me that the intention instead was to SUBTRACT this width
when setting the size of the layoutContainer. I also changed the code
to take into account the padding and to check whether the
layoutContainer uses boxSizing set to border-box, in which case we don't
need to subtract anything.

The result is that now the contents are properly smaller than the div
that surrounds it, and it no longer causes a scroll bar to appear.