WT-11301: Add test case for clean client server shutdown
Add a test case that will shutdown the server with an outstanding
request to a static resource running on a thread.

If the resource is destructed before the server is shutdown,
this can result in a use after free when the resource's request
handler accesses the destructed resource object, e.g. to create a
continuation.