WT-7279: Added WEmailEdit and WEmailValidator
WEmailEdit implements an <input type="email">. We're adding support for
the following attributes:

- multiple
- pattern

Note: I chose not to make WEmailEdit a WLineEdit, since there's no
obligation for browsers to display a WEmailEdit as a WLineEdit. Similar
functionality can be attained with a WLineEdit by adding a
WEmailValidator to a WLineEdit.

The WEmailEdit does sanitization in accordance to the WHATWG
specification: removing newlines and carriage return characters (if not
set to multiple), and removing trailing and leading whitespace.

WEmailValidator does validation in accordance to the WHATWG
specification: the email addresses accepted by WEmailValidator are a
subset of all possible email addresses, but we chose to go with the
standard that is followed by web browsers.

Co-authored-by: Ruben Kindt