WT-11348: Clearly and unambiguously define pathInfo()/scriptName()/etc.
The semantics of pathInfo() and scriptName() used to depend on when they
were called.

Before the entrypoint was matched:

- scriptName() is equivalent to the SCRIPT_NAME CGI parameter
- pathInfo() is equivalent to the PATH_INFO CGI parameter

After the entrypoint matching:

- scriptName() is the SCRIPT_NAME + the entry point
- pathInfo() is the PATH_INFO without the entry point

e.g. if SCRIPT_NAME is /one.wt, PATH_INFO is /two/three and the
entrypoint is /two:

Before entrypoint matching:

- scriptName() returns /one
- pathInfo() returns /two/three

After the entrypoint was matched:

- scriptName() returns /one/two
- pathInfo() returns /three

This makes things confusing and leads to issues, in fact:

- for ISAPI it appears that the string pointed to by the result of
  persistentEnvValue("PATH_INFO") is modified. removing the entry point,
  so the next call would return a different result!
- for FastCGI there's an if (scriptName_.empty()), so its result will be
  fixed to be str(envValue("SCRIPT_NAME")) after the first call, and not
  be updated after the entrypoint was matched.
- for FastCGI there's a similar issue with the if (pathInfo_.empty())

Another issue we're incidentally fixing is the fact that since we can
now have dynamic parts in entrypoint paths the length of the part
matching the entrypoint and the length of the entrypoint's path may
not be the same (in fact it most likely won't be).

We now have:

- WebRequest::scriptName(), which ALWAYS returns SCRIPT_NAME (empty
string for wthttp)
- WebRequest::pathInfo() which ALWAYS returns PATH_INFO (full request
path for wthttp)
- WebRequest::entryPointPath(), which is the entrypoint part of
pathInfo()
- WebRequest::fullEntryPointPath(), which is SCRIPT_NAME +
entryPointPath()
- WebRequest::extraPathInfo(), which is pathInfo() without the
entrypoint

I also added plenty of internal documentation, so that we hopefully have
no more confusion in the future.

To make this happen, http::Request was adjusted so that request_path is
always the full_path, and request_extra_path was replaced by
extra_start_index.