WT-11348: fix entry point matching with script name
When we were using FastCGI or ISAPI, we would fail to properly match
some entrypoints. If scriptName is /hello.wt, and we have an entrypoint
deployed at / and at /entrypoint, then a request for
/hello.wt/entrypoint would match / instead of /entrypoint.

This issue was fixed by both matching with and without the script name
and comparing the two matches according to the following criteria:

- Successful matches are preferred over unsuccessful matches
- The most static or least dynamic path is preferred, e.g.:
  /one/${two}/three is more dynamic than /one/two/${three}, since the
  dynamic part of the second string comes later.
- Finally, we prefer the deepest match. This comes down to the
  entrypoint with the most slashes.

EntryPointMatch::operator< was implemented to reflect those criteria:
if match1 < match2, that means that match1 comes before match2, and is
this preferred over match2. If both matches are equivalent we prefer the
match without script name.

Also, as per the CGI standard, both SCRIPT_NAME and PATH_INFO can be the
empty string or must start with a slash. We'll make sure that we don't
fail some sort of assertion by simply checking this condition and
logging an error otherwise.