WT-11348: make Configuration testable
We need the #include "Wt/WConfig.h" because we use stuff like
WT_THREADED. If some files using the Configuration include it,
and others don't, then there will be an ABI mismatch, leading to
all sorts of weird crashes and errors.

We need to allow the configurationFile_ to be empty, and then just
make it skip over readConfiguration() If we want to be able to make
isolated instances of Configuration.