WT-11384: fix Wt::Mail::Client throwing on connect
If the resolver failed, we would throw an exception, instead of logging
an error and returning, as documented.

Also, the implementation of `tcp::resolver::resolve` we were using with
the `tcp::resolver::query` argument is deprecated, so we're now using
the version that takes `host` and `service` (i.e. port in our case).

We'll capture any possible error from the resolver, and log an
appropriate message.

I also made sure to log the error message when we fail to connect.