WStackedWidget: corrected setCurrentIndex() documentation, added test
The documentation of WStackedWidget contained the broken line
"The default value for current index is 0 (provided thath".
I think what it meant to say is that it's -1 if there are no child
widgets, and 0 when you add at least one widget.

I also added a test case to ensure that the documentation is indeed
correct.