WT-11376: fix FastCGI setRedirect
When entering the widget gallery in progressive bootstrap mode, with
some web servers like Caddy, the response body would be:

Content-Type: text/html; charset=UTF-8
Status: 302

The reason for that was the double \r\n, which causes the web server to
treat the rest of the response as the response body, rather than
headers. It appears other web servers were more forgivable and would
simply wait until after Status: to consider the headers sent, but Caddy
is more strict.