WT-11301: add WServer::addResource taking a shared_ptr
This deprecates the version of addResource that takes a raw pointer.
Using the shared_ptr we can enforce that resources outlive the server
instance, so requests are not at risk of being routed to a deleted
resource.

All of the examples and tests have been changed to use the new version
of addResource.