WT-11424: fix WDialog::setClosable(false) "widget not in container error"
If a dialog was already not closeable, and we called setClosable(false),
then there would be an error because we would try to remove closeIcon_
(a null pointer) from the title bar.

By checking early on whether or not the dialog is already closable or
not we won't get that error any more.

We can also remove the if (!closeIcon_) since it is now already
implied by the earlier check. (Whether or not a dialog is closable
is implicitly encoded in the presence of the closeIcon_.)