WT-10908: fix WSuggestionPopup current selection updating with BS5
One of the BS5 checks failed, since it was looking for APP.theme instead
of WT.theme. I refactored all of the different checks that check for
BS4 or BS5 to a single isBS5() function (we're not checking for BS4,
since we decided not to add WBootstrap4Theme).

Instead of using className, I opted for the more robust approach of
using classList.remove and classList.add. I also made sure that the
removal/addition of the active class on the <li> is nicely paired with
the removal/addition of the active class on the <a> (firstChild).

As a little extra, I added a scrollToSelected(...) to the arrow key
selection code, so that the currently active item is automatically
scrolled into view.