WT-11433: enable drop location for the folder view
This used to work because FolderView overrides onDropEvent and Wt
generated the event even if it wasn't enabled (leaving it up to the view
to check if drops were enabled). Issue 8636 changed it so the event is no
longer generated if drops are not enabled. Therefore the example now needs
to explicitly enable drops.