WT-11423: empty internalPath() for static resources
We could get a warning about "adding '/' to start of internal path"
when a resource was added like:

addResource(resource, "path");

If --deploy-path was /deploy-path, this would deploy the resource at
/deploy-path/path, but call WResource::setInternalPath("path");

Since the concept of an internal path doesn't really make much sense
in the context of static resources (internal to what?), we decided to
remove the setInternalPath(...) call, so that the internal path is
empty.

WResource::url() still returns the last path that the resource was
deployed at, and I added some documentation to clarify that.