WLocalDateTime.C: fix compilation with std::chrono
When building Wt with -DWT_CPP20_DATE_TZ_IMPLEMENTATION=std
in a Fedora 38 container with GCC 13, this code failed to
build. The date library is not included, and we'll need to
use cpp20::date instead, which corresponds to ::date when
using the date library, and std::chrono when using we're
using the standard library.