wthttp: modernize httpDate
Instead of relying on gmtime we can now just use the date library
or the C++20 version of std::chrono, eliminating our need to rely
on platform-specific code.

We also now output the day of month with a leading '0', since this
appears to be the way it's done in general.

I moved it to libwt and put it in web/DateUtils.h/C, so that we
can write tests for it (and maybe use it in other places).

StaticReply.C and FileUtils.C were also updated as a result of changing
the type that httpDate() takes to a time_point.