WT-11270: implement somewhat proper FindHPDF module
I named the module FindHPDF rather than FindHaru or FindLibHaru to
reflect the name of the lib file, headers, and prefix of functions.

I did keep the HARU_DYNAMIC, ENABLE_HARU, HARU_PREFIX, and HAVE_HARU
variables the same, so that the CMake arguments that need to be provided
to Wt are more or less the same. It's only if the user wants to specify
a specific include dir or library that they will need to provide
HPDF_INCLUDE_DIR/HPDF_LIBRARY instead of the old
HARU_INCLUDE_DIR/HARU_LIB.

Depending on whether we're using Haru 2.3.0 or later versions, the names
of the libraries are different.

On Windows, it was:
- debug/dynamic: libhpdfd
- debug/static: libhpdfsd
- release/dynamic: libhpdf
- release/static: libhpdfs
On Linux, it was:
- dynamic: hpdf
- static: hpdf or hpdfs depending on how it was built
  (using autoconf or CMake)

Now, it's always hpdf.

We're now using the imported HPDF::HPDF target. This means we no
longer need the legacy add_definitions(-DHPDF_DLL), but we can now
use target_compile_definitions(HPDF::HPDF INTERFACE HPDF_DLL) to make
sure anything that depends on libHaru adds HPDF_DLL.

The examples/CMakeLists.txt assumes that we're using our own
binary distribution on Windows (with winstng). We still patch
Haru 2.4.3 to add the d suffix to the debug library, so it's
hpdfd.lib/dll for the debug build, and hpdf.lib/dll for other builds.