Fix JWt translation in OAuthService
std::make_shared<Impl::RedirectEndpoint> ended up being translated to
new RedirectEndpoint(...), which wouldn't compile, since
RedirectEndpoint is a static inner class of the static inner class Impl.

Writing it this way does translate properly. This appears to be the
easiest way to fix it, rather than trying to fix the issue in CNOR.