WT-11578: Dbo Forms: Correctly validate invalid dates
Say that your application expects dates in the `yyyy-MM-dd` format. If
the user would enter them in the `dd/MM/yyyy` format, the application
would clear the date in the field and as long as the date field isn't
required, validation would succeed. This is due to WDate::toString
returning an empty string for invalid date objects. This empty string
would then be passed to the WDateValidator via the validate functions of
WFormModel.

The Dbo Forms are now aware of this and will set validation result as
the value for the field in the model when the input cannot be converted to a
valid date object. Our customized `Wt::Form::Dbo::FormModel` already has
support for this and will check in its `validateField` function if the
value in the model is of type `Wt::WValidator::Result` and display the
correct "error".