WT-11578: Dbo Forms: Correctly validate invalid times
Similar as with the specialized WFormDelegate for WDate objects: if the
user enters a time object in an incorrect format, the application would
clear the field and as long as the time field isn't required, validation
would succeed. This is due to WTime::toString returning an empty string
for invalid time objects. This empty string would then be passed to the
WTimeValidator via the validate functions of WFormModel.

The Dbo Forms are now aware of this and will set the validation result
as the value for the field in the model when the input cannot be
converted to a valid time object. Our customized `Wt::Form::Dbo::FormModel`
already has support for this and will check in its `validateField` function if
the value in the model is of type `Wt::WValidator::Result` and display the
correct "error".