WT-10485: Do not include `GL/gl.h` on MacOS
The implementation on Apple systems is called `OpenGL`, not just `GL`.
As such, the inclusion of `GL/gl.h` will not succeed on MacOS. This
inclusion should be hidden behind a check that ensures we are not on an
Apple system.

For Apple systems the inclusion `OpenGL/OpenGL.h` will take the place of
the one that was put behind the check.