WT-10485: Do no longer rely on the included GLEW source
Since CMake 3.1.0 there exists a `FindGLEW`. That way we can simplify
the `CMakeLists.txt` somewhat.

If `OpenGL` is included already, `GLEW` will be required. If it cannot
be found, an error will be shown to the user, telling them `GLEW` cannot
be found automatically in their predefined paths. If this happens, they
are prompted to define `GLEW_ROOT`, which points to their custom GLEW
installation.

Since `GLEW` is now no longer bundled with Wt, the license has been
removed as well.

This commit will pave the way to also use `FindOpenGL`, instead of the
custom `WtFindGL.txt`.