WT-10485: Update license since GLEW is no longer included, only external
We no longer budle GLEW with Wt, but rely on a custom installation, that
can be found with either CMake's `FindGLEW` or by setting the
`GLEW_ROOT` CMake argument.

Aside from now only relying on GLEW being external, the URL for GLEW was
also updated. The sourceforge link is outdated, and points to an older
release (2.1.0). Github has the latest version (2.2.0).