WT-10512: Create test for the `setDisabled()` of `WCompositeWidget`
This widget holds more complex widgets, often widgets that hold a lot of
functionality, and are comprised of a lot of items in the DOM tree.

When a widget is disabled, the disabled state of that widget is set, and
the state is propagated to its children. The child's state is only
changed visually, their `isDisabled()` will return `false`, unless it
was set to `true` before.

These tests ensure that this functionality remains, and the following
commits do not change them.