WT-10512: Propagate the disabled state to new DOM elements
Whenever a DOM element is added to the tree, when its parent has been
disabled, it should also be disabled. Note that this does NOT use
`setDisabled()`. This is purely used to manage the widget's own disabled
state, regardless of any of their parent or siblings in the DOM tree.