WT-7645: Add an empty href attribute to `WAnchor`s on JavaScript update
When a `WAnchor` is created this can be done through progressive
bootstrap, meaning that an initial page is sent as HTML, with subsequent
JavaScript calls for updates to the page. In this case, if a `WAnchor`
points to an empty link (used for navigation as part of a widget), the
"href" attribute was removed. This caused the anchor not to be part of
the tab chain of the page. This chain can be used to traverse the
functional items of a webpage with the keyboard alone. Going over all
elements that can receive focus (like anchors, buttons, ...).

It is now ensured that if we are indeed loading the item via JavaScript,
that the "href" is not removed, and the item remains part of the chain.

The original ticket does mention that a selection in such a menu / tab
breaks a more global listener, but that is simply part of the focus
change inherent with selecting an item.