WT-11741: Correct the jWt compiler directives for a mutex unlock
Under `WebSession::Handler::unlock()` the resulting translated
Java code was empty. This was because instead of the compiler directive
ensuring that there was an implementation for Java, it omitted any code.
Now the right code is called, which has a Java counterpart.

This function is similar to `WebSession::Handler::release()`, with the
exception that it only performs the unlock. Any of the queued events
will be omitted.