WT-11804: Add edge cases to the tests for the `WEmailValidator`
There is a case where it is allowed to have special characters in the
name (like a hyphen). Likewise a hyphen is also the only special
character allowed in the domain part of the email address.

This does not test the client-side validator, where the actual issue
occurred. For now, in our current set-up this isn't something we can
easily integrate.