WT-9596: Fix regression in Postgres backend float/double precision
A fix performed in WT-9490 ensured that subnormals can be retrieved
correctly from the Postgres backend. In doing to, `std::stof` and
`std::stod` were replaced with `boost::spirit`. However, this parser has
the chance to lose some precision:
https://github.com/boostorg/spirit/issues/668

To ensure that we keep the floats and doubles as precise as possible, we
will return to using `std::stof` and `std::stod`. If we then encounter
the `std::out_of_range` exception, we fall back to `boost::spirit`.

This does mean that the conversion is again locale dependent. In a
following commit we will ensure that we use `std::to_chars` and
`std::from_chars`, from `<charconv>` to ensure that Wt performs
correctly regardless of global locale.

Note that not all tests are equal for all backends. Notably, MariaDB has
different syntax for float casting, and seems to have some precision
conversion edge-cases, much like SQLite3. This comes down to both of
them having less significant decimal digits than Postgres.