WT-9595: Rely on `charconv` for float/double precision for Postgres
The Postgres backend used to rely on `std::stof` and `std::stod`, which
was replaced with `boost::spirit`. The latter, however, loses some
precision. Hence we moved back to the original `std` functions, with the
boost library being the fallback in case we encounter a
`std::out_of_range` exception.

Now, we use `std::to_chars` and `std::from_chars`, to parse and create
strings from floats/doubles. This has the advantage that is does not
rely on an external library, nor is it locale dependent.

The caveat here is that this is C++17 (or C++14 as an extension).

Again note some specific tests have certain cases for some backends.
This is because MS SQL Server does not have the concepts of NaN or
infinity. Nor is MariaDb able to hold NaN.