WT-11302: Avoid use-after-free with closed listeners
When we call handleResume() or handleStop() all acceptors are closed
and all listeners deleted. However, closing the acceptor causes
handleTcpAccept or handleSslAccept to be called with an error code.

Previously, this would mean that the listener argument to these
functions would be a dangling pointer and dereferencing it is undefined
behavior. By letting the callback take a weak_ptr we can ensure that
we don't dereference a dangling pointer.