WT-11408: `WebController::running_` concurrent access
There are two methods that can set this state, namely `start()` and
`shutdown()`. However, in case of the latter, while it is writing, it is
possible that either `handleRequest()` or `requestDataReceived()` do a
concurrent check on this variable.

This boolean was thus converted to a `std::atomic`, to ensure that these
writes and reads cannot lead to race issues. There was also the option
to rely on the mutex, but since this mutex is more used for session
management, the `std::atomic` approach seemed more appropriate.