WT-11408: Guard possible concurrent access to `Wt::Http::Client::impl_`
Both the `abort()` and `emitDone()` call use the `impl_` pointer. The
former requests a lock, while the latter resets the pointer. Since it is
possible that they are called concurrently, this may result in undefined
behaviour.

While this uses a `std::weak_ptr`, making sure the reference count to
the pointer is atomic, this does not make the actual operations
happening on the pointer atomic.

Note: this also removes a lot of the `Wt::Signals::Impl::ProtoSignal`
warnings that `ThreadSanitizer` reported in the tests. This is simply
due to how the tests are implemented.