WT-11408: `Connection::waitingResponse_` concurrent access
There are two methods that can set this state, namely `handleReadBody()`
and `handleWriteResponse0()`. When calling `waitingResponse()` to read
the value, it is possible that a concurrent check on this variable
occurs.

This boolean was thus converted to a `std::atomic`, to ensure that these
writes and reads cannot lead to race issues.