WT-11848: Solve an uncovered case in the `WSuggestionPopup` javascript
There was the possibility that the value would be undefined. In case
that `setDropDownIconUnfiltered()` was also set to `true`, this resulted
in the resulting `text` variable being undefined, instead of being equal
to an empty string.

Instead of correcting this in the C++ code, and calling it with an empty
parameter, we ensure that this value can not be undefined when the popup
is shown via `showAt()`.