WT-11852: Rename `localized` parameter to `localizedString`
This is part of `WDateTime::toString()`, allowing the resulting formatted
string to be fully localized. This means for example that "ddd" results
in "Mon" in English, on a Monday, but in "Lun" in Spanish. This will
depend on the locale of the `WApplication`.

The parameter was changed from `localized` to `localizedString` to
unambiguously indicate that the resulting string would be localized, not
the value of the time or date.