WT-11852: Never localize the `Mail::Message`'s date header
The date header is now RFC-5322 compliant. Before, if a localized day
(abbreviation) or month (abbreviation) was specified, this string was
picked to generate the date header for the email.

According to RFC-5322 (https://www.rfc-editor.org/rfc/rfc5322#page-14)
this header only accepts English names for the days and months.