WT-11882: Add flex direction column to `fieldset`
With Bootstrap 5 specifically, the content of a  `WGroupBox` was not
correctly displayed. Putting `display: flex` on the `fieldset` element,
would cause its legend to take the width of the entire view, and put all
other widgets that are contained in the `div` below it, on its right.

Specifying the flex direction ensures that the div below it, is actually
rendered below the legend.