WT-11890: Ensure the hidden() signal cannot result in a use-after-free
If a user were to delete the popup after it is hidden, which is allowed
according to the documentation, the `isRendered()` check in
`WPopupWidget::setHidden()` will try to access a deleted object.

See:
https://www.webtoolkit.eu/wt/doc/reference/html/classWt_1_1WPopupWidget.html#details

The signal for `hidden()`or `shown()` is now only emitted after all
other checks have happened, so no use-after-free occurs.