WT-11984: Fix tableview drag & drop bug 2
When a tableview is configured for dropping below lines and when
dragging an item below the table, the tableview behaves as if you want
to add the item on the bottom, which is fine. But when an item was
dragged right or left from the table, the whole drag & drop process
crashed. This commit fixes this. Dragging left/right now behaves as if
the mouse is still inside the table.

A small imperfection that remains is that the mime icon only stays with
the mouse when leaving the table to the left or right but not when
leaving the table bottom/top. I do not deem it worth it to go further
looking to solve this at this time.