WT-10512: Disable newly add `WTreeNode`s
Although these items are already unselectable, they still have to be
disabled. Otherwise it would look like they can be selected. This is
because the "Wt-disabled" class marks them visually as disabled.

A check is added to `WCompositeWidget::isDisabled()` to ensure the
implementation exists. A widget may on construction perform this check
(like WTree, which updates its children), without the implementation yet
existing.