WT-11932: Allow immediately replacing `WPushButton`s `WPopupMenu`
This menu's lifetime is managed by the `WPushButton` itself, moving it
to a member variable. When one popup is already assigned to the
`unique_ptr`, assigning it again should remove the old button and move
the new pointer in place. However, since the `WPopupMenu` can be deleted
before the button is, the destructor of the `WPopupMenu` sets the
button's menu to `nullptr` on destruction, if it has a button.

We now check that if the `WPopupMenu` has a button, that if the button's
menu is the current destructing `WPopupMenu`, that the button's menu can
be reset to `nullptr`. Otherwise we do not touch it. That ensures that
after the `std::move` of the new `WPopupMenu` has taken place, it will
not be reset.