WT-7196: Remove deprecated Pango call
The Pango method `pango_fc_font_lock_face()` has been deprecated for a
long time now, and should really be removed in case they ever remove the
function.

From their documentation it states that it should be replaced with:
 - `auto hbFont = pango_font_get_hb_font(font)`
 - `hb_ft_font_lock_face(hbFont)`

This however is not valid, since it returns `null`. This is due to the
way the `hbFont` is constructed. It is a generic `hb_font`, not a
specific `hb_ft_font`. This causes the `FT_Face` to be missing.
Constructing a `hb-ft_font` requires an `FT_Face`, sending us in a loop.

We thus require a different way to get the name of the face file. We do
this by avoiding `FT_Face`, and by looking in fontconfig's object for
the filename.