WT-11772: Resolve some `Wmisleading-indentation` warnings
This warning occurs when indentation may make it seem like code is
executed is a certain branch, but it is actually excluded.

Example:
```
if (condition)
  expression1;expression2;

---

if (condition)
  expression1;
expression2;
```

In both cases `expression2` is not part of the `if` branch.

The warnings in the sqlite3 backend amalgamated code were simply
suppressed, since this is copied code.