WT-11772: Resolve some `Wunused-variable` warnings
This warning occurs when a variable is declared, but never accessed.

Note that some instances of these warnings are still in the code. In
some cases this is because the reference of the variable is accessed,
and it is caught as an unused case, whereas it isn't *really*.

In some other cases, this is a complaint about the `LOGGER("classname")`
macro. This creates a logger instance of the right scope. Other macros
can then be used (like `LOG_WARN`) for the specific type.