WT-11772: Resolve some `Wsign-compare` warnings
This warning occurs when a comparison is performed between two types
that cannot be safely compared. Most often this happens if a comparison
between a signed and unsigned type take place.

The easiest solution is to cast the signed value to unsigned. In some
places the size is "guaranteed" to be positive, and thus can be casted
to a signed value.