WT-11772: Use `WT_MAYBE_UNUSED` for the logger
This is used for `LOGGER` instances. Where we use macro's to call the
logging functionality. This isn't caught by the compiler, which warns
that the logger is not used.

Nearly all files have a logger associated with them, this can be unused,
if the logger is not written to in the code explicitly. We declare it as
`[[maybe_unused]]` to avoid compiler warnings.