WT-11974: Override createFormWidget in FormView
We noticed a bug in the FormView logic where making a field dynamically
visible would result in the form only showing the label of the field,
but not the widget. This is because making a field invisible will remove
it from the widget map kept by WTemplate. Making it visible again after
a while will result in resolveWidget to return a nullptr. The widget
will then need to be recreated by using
WTemplateFormView::createFormWidget. The default implementation of this
function simply returns a std::unique_ptr<WWidget> and has no effect. We
need to override this function in the Dbo Forms to ensure that the
correct widget from the delegate is used.