WT-11985: `WStringListModel` drag/drop in `WTableView` clears content
When a `WTableView` is used, and it allows dragging and dropping its
items, the content of the `WStringListModel` is not correctly updated.

This comes from the fact that this model contains a fix for its editable
fields, where this content is set to be the content to be displayed in
the view. This fix interfered with the normal working of the model, and
caused the content for the display role to be cleared.