WT-11934: Remove the `WSlider` vertical orientation fix
This fix ensured that we apply the transformation fix, where the widget
would be resized in a manner behaving more like an image, thanks to the
CSS transform rotation, yielding weird dimensions, and scaling. The
widget is "transformed" such that a box around it is drawn, allowing it
to scale to its parent container.

However the previous changes to the `WSlider` now make this change
obsolete.