WT-11934: Set the `WSlider`s native control's orientation property
This property indicates the orientation of the widget. This should be
done with this class for proper native support, instead of rotating the
widget.

For Firefox we can use the `orient=vertical` property.

Whereas for Chrome and Safari we rely on `-webkit-appearance:
slider-vertical`.

The agent version check was removed since all of these agents are very
outdated. We already dropped nominal support for Internet Explorer, so
any of these user agents should barely occur (since most of them
auto-update). Additionally, this check should have been performed on the
setter, not the getter.