WT-11934: Offset the `WSlider` handle correctly in Wt default themes
Wt comes with two themes, "default" and "polished". The slider handle in
the default theme is never positioned correctly. This is because in the
"polished" theme, the handle can be positioned in the middle, whereas
for the "default" theme it should be positioned next to the slider bar.

For the "default" theme we now calculate an offset, based on the
widget's width or height (depending on orientation).