WT-11934: Allow the `WSlider` tick length to be set
This dictates the length of a single tick, which is either the line's
width or height, for a vertical or horizontal slider respectively.

If this remains on auto (which is currently default behaviour), the tick
lines way grow to take up as much room as the slider itself.

While the documentation states that it ought to be resized, I feel it's
safer to also give the user control over this. That way; even if they
don't resize, they may set the length to arrive at a slider that
visually is what a user would expect.