WT-12012: Call the `updateDom` of a `WTreeViewNode`s parent
The `updateDom` function was added to `WTreeViewNode` to ensure that
updating the item would result in all its components being correctly
disabled, if the widget was disabled.

However, this was missing the call to its parent. This call set all the
interaction events with the keyboard and mouse. Hence this was missing
several possible event.