WT-12005: Add steps to the `WSlider`
This facilitate the slider to be moved by a certain step value. The step
must be a positive value, but it isn't necessary that the slider can be
neatly divided by the step.

For example if a slider with min = 0, max = 50, step 7 exists, it will
be able to go up until the value of 49, never reaching the max. It is up
to the developer implementing the `WSlider` to choose value that make
sense.