WT-12005: Correct the native `WSlider` to update its step and value
The native slider functions by setting the value on changes. This value
needs to be between the min and max value, and can be incremented from
the minimum by a (or multiple) value(s) of step.

By default this value is 1.

This commit forces the `WSlider` to call its `updateDom` function, which
before didn't happen, since it was a purely native element, and didn't
receive any changes.

On dragging the handle of the slider it will also only go over the
correct step values if this is enabled. Note that upon fast mouse
acceleration the handle may become out of line with the mouse icon, in
which case the mouse icon represents the correct placement.