WT-12022 (and WT-11669): Collect invisible changed before stateless slot
When collection all JavaScript changes both on initial load (with the Wt
framework, `serveMainAjax`) and on subsequent requests
(`serveJavaScriptUpdate`), first all visible changes are added, then
invisible changes. It used to be that the prelearned slots of stateless
slots would be considered as part of the visible changes. However, this
caused some changes to invisible widgets to be added to the stateless
slots. While often not really an issue (though considerably wrong), it
can be an issue if the content of the page changes.

Suppose you have a text widget that is invisible, and its content is
changed to "some text". This text widget is displayed on a page that
also contains a menu. When this change is executed the id can be found
on the page, since it is simply not visible, but still part of the DOM.
This change is also added to the first stateless slot the application
encounters, say of a menu item.

However, the widget is later removed, invalidating the id. When the
handler of the menu is executed (on click), it still has the now
non-existent id, which will throw a JavaScript error. Since it resolved
to "null", and no function can be executed on it.